backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
beanie==1.25.0
pymongo>=4.6,<4.8
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openpyxl==3.1.2
Pillow==10.3.0
rembg==2.0.56
imgkit==1.2.3
Jinja2==3.1.4
qrcode[pil]==7.4.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
apscheduler==3.10.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
